+++ common/core/src/main/java/is/codion/common/state/DefaultState.java
@@ -113,11 +113,6 @@ final class DefaultState implements State {
   }
 
   @Override
-  public void accept(Boolean data) {
-    set(data);
-  }
-
-  @Override
   public boolean addListener(Runnable listener) {
     return observer().addListener(listener);
   }