+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 ==================
 
 ## 0.17.27-SNAPSHOT
+### is.codion.common.core
+- Value, default implementations added for accept() and map().
 ### is.codion.framework.domain
 - DefaultEntityDefinition.validatePrimaryKeyAttributes(), key column indexes now sorted before being displayed in error message. 
 ### is.codion.framework.db.local